Revert "Revert "Terrian Tesselation ongoing""
This reverts commit 8507da0646e50aba8dc4f7f3c795ba484b011307.

+++ Object/TestCube.cpp
@@ -9,7 +9,7 @@ TestCube::TestCube()
 	{
 
 	for (int i = 0; i < 6; i++) {
-		D3DXMatrixTranslation(&world[i], 128 + 10 + 4*(i%3) , 0, 128 + 10 * (i/3));
+		D3DXMatrixTranslation(&world[i], 128 + 10 + 4* (float)(i%3) , 0, 128 + 10 * (float)(i/3));
 		wBuffers[i] = new WorldBuffer();
 		wBuffers[i]->SetWorld(world[i]);
 	}
